refactor(attribution): extract AttributionCapability into its own module
