+++ src/capabilities/mod.rs
@@ -5,6 +5,7 @@
 
 pub(crate) mod approval;
 pub(crate) mod ast_grep;
+pub(crate) mod attribution;
 pub(crate) mod background;
 pub(crate) mod client_commands;
 pub(crate) mod config;
@@ -22,6 +23,7 @@ pub(crate) mod your;
 pub(crate) use crate::goal::GOAL_CAPABILITY_ID;
 pub(crate) use approval::{APPROVAL_CAPABILITY_ID, ApprovalCapability};
 pub(crate) use ast_grep::{AST_GREP_CAPABILITY_ID, AstGrepCapability};
+pub(crate) use attribution::{ATTRIBUTION_CAPABILITY_ID, AttributionCapability};
 pub(crate) use background::{
     AgentRunResult, AgentSpawner, BACKGROUND_CAPABILITY_ID, BackgroundCapability,
     BackgroundRegistry,
@@ -31,9 +33,8 @@ pub(crate) use config::{CONFIG_CAPABILITY_ID, ConfigCapability};
 pub(crate) use goal::GoalCapability;
 pub(crate) use hooks::{HOOKS_CAPABILITY_ID, HooksCapability};
 pub(crate) use host::{
-    ATTRIBUTION_CAPABILITY_ID, AttributionCapability, CodingBashCapability,
-    CodingCliEnvironmentCapability, ENVIRONMENT_CONTEXT_CAPABILITY_ID, SETUP_CAPABILITY_ID,
-    SetupCapability,
+    CodingBashCapability, CodingCliEnvironmentCapability, ENVIRONMENT_CONTEXT_CAPABILITY_ID,
+    SETUP_CAPABILITY_ID, SetupCapability,
 };
 pub(crate) use repo_map::{REPO_MAP_CAPABILITY_ID, RepoMapCapability};
 pub(crate) use worktree_cmd::WorktreeCapability;
